Moved telemetry related apis in telemetry blueprint from app.py
